BattleJack release
Implements a DM game called BattleJack, a blackjack variant played 1v1 by users. The goal is to get as close to 21 without going over. This commit adds the entire game, the ability to start a game against another player, to hit, to stand, to win or lose. Front and back end have been added.